+++ src/components/Estatisticas/Estatisticas.jsx
@@ -1,7 +1,7 @@
 import React, { Component, lazy, Suspense } from "react";
 import Menu from "../MenuSuperior/Menu";
 import Auxiliar from "../MenuSuperior/Auxiliar";
-import Rodape from "../Rodape/Rodape";
+import { Rodape } from "../Rodape";
 import NullView from "./NullView";
 
 import {
